// src/types/components.ts
export type ComponentNames =
  | 'CPU'
  | 'GPU'
  | 'RAM'
  | 'Storage'
  | 'SSD'
  | 'HDD'
  | 'Motherboard'
  | 'Power Supply'
  | 'Sound Card'
  | 'Cooling'
  | 'Case Fan'
  | 'Cases'
  | 'Mouse'
  | 'Keyboard'
  | 'Headset'
  | 'Microphone'
  | 'Webcam'
  | 'Monitor'
  | 'Optical Drive'
  | 'Cable'
  | 'Laptop'
  | 'Printer';

export type ComponentSlugNames =
  | 'cpu'
  | 'gpu'
  | 'ram'
  | 'storage'
  | 'ssd'
  | 'hdd'
  | 'motherboard'
  | 'power'
  | 'soundcard'
  | 'cooling'
  | 'casefan'
  | 'case'
  | 'mouse'
  | 'keyboard'
  | 'headset'
  | 'microphone'
  | 'webcam'
  | 'monitor'
  | 'opticaldrive'
  | 'cable'
  | 'laptop'
  | 'printer';